feat: duration

+++ packages/nocodb-sdk/src/lib/sqlUi/MssqlUi.ts
@@ -1144,7 +1144,7 @@ export class MssqlUi {
         colProp.dt = 'double';
         break;
       case 'Duration':
-        colProp.dt = 'int';
+        colProp.dt = 'decimal';
         break;
       case 'Rating':
         colProp.dt = 'int';
